불합리 수정
1. 오버플로우 수정
2. Param에 더미데이터 적재
3. 미사용 함수 주석처리

+++ CRUX_GUI_Renewal/CRUX_Renewal/Class/ServerInterface.cs
@@ -181,7 +181,7 @@ namespace CRUX_Renewal.Class
                 {
                     Marshal.Copy(pSndMsgBuff, cMsgBuf, 0, (int)nMsgSize);
                 }
-                if ((int)pSndMsgBuff != 0)
+                if ((long)pSndMsgBuff != 0)
                     Marshal.FreeHGlobal(pSndMsgBuff);
             }
